Moved validate functions as agreed with agnes
